Static Map Edit Console Logs added for debugging

+++ model/src/main/java/com/kurabiye/kutd/model/Enemy/Enemy.java
@@ -102,7 +102,7 @@ public class Enemy implements IEnemy {
         }
     }
 
-    public synchronized void move(long deltaTime){
+    public synchronized void move(double deltaTime){
 
         
         if(enemyState == EnemyState.DEAD || enemyState == EnemyState.ARRIVED) {
