+++ model/src/main/java/com/kurabiye/kutd/model/Listeners/IGameUpdateListener.java
@@ -8,7 +8,7 @@ public interface IGameUpdateListener {
      *
      */
 
-    void onGameUpdate(float deltaTime); // Method to be called on game update with delta time as parameter
+    void onGameUpdate(double deltaTime); // Method to be called on game update with delta time as parameter
     // This method will be implemented by classes that want to listen for game updates
 
 }
